test: extract delegated skill evidence suite
